Resolving Punctuation Errors in Coalition Jobs (#2193)
* Remove Comma Errors

http://www.dailywritingtips.com/punctuation-mistakes-1-unnecessary-commas/

It is an error to separate a subject from a verb.

* Peacekeeper Commas

"I need a car, to get food" should be "I need a car to get food". http://www.dailywritingtips.com/punctuation-mistakes-1-unnecessary-commas/
#3 and #5 apply here.

* Dangling Modifier Fix (and consistency)

https://owl.english.purdue.edu/owl/resource/597/01/
previously, "as part of the program" was modifying "transport", which is incorrect; there is no clear subject to the introductory phrase "as part of the program". It now has a subject (the volunteer settles).

Also, look at Coalition: Transfer 1 just directly below the modified missions; they are done correctly by default.

* Dangling modifier and consistency

Just look at Arach 1; that's how grammar should be done.

* Fast Courier extra comma

"I need a car, by tuesday" --> "I need a car by tuesday".