Fix errors/inconsistencies in jobs.txt
 - hai jobs: dialog strings contained improperly escaped doublequotes. Fixed by using `` instead of "", as is done with other jobs/dialogs that have embedded doublequotes.
 - syndicate world jobs: Updated the "Corporate retreat" mission displayname to avoid confusion
 - wanderer jobs: Updated the "Invasive Species" missions to avoid confusion

Use of <planet> in the display name does not load the name of the intermediate destination that the planet has to visit, but is the same as <origin>.
Using <stopovers> would be better, but this text replacement hotword also gets the system name, so it will overflow the "Missions available" panel. It also may have more than one value.