Fix to changing a system's music in an event
At the moment, Engine::EnterSystem plays music before IncrementDate is
run (which manages event handling). This means that if an event changes
the music of the system you're in, you won't get the change until a day
later. This fixes that by reordering the code. Should also fix the same
issue with haze, if an event changes that.