Increase maximum volume back to 100%
This increases the upper limit on what the volume can be set to back
to 100% (as before 006238e85b7c9647e011c1d6bebcee40216c81ee).

The default volume remains unchanged (now 12.5% of maximum).

Volume in existing configurations remains unchanged. In the
configuration file, the volume value now has the range from 0 to
4 (instead of 0 to 1).